requirements.txt
requests==2.20.0
Flask==1.1.1
lxml==4.3.3
Telethon==1.11.3
selenium==3.141.0
Flask_Script==2.0.6
PySocks==1.7.1
